Fixed output of several files at once
The output parameter contains an array of strings, where each string corresponds to an input item. When converting the array to a string, a comma is inserted by default as delimiter, which corrupts the resulting CSS. This commit fixes the issue by using the Array.prototype.join(delimiter) function and providing a pure array to grunt.file.write.